Updated update error case
Now, we can update just the description of a sport. We can save an update with a no different name, but we cannot save a name already exists on DB 

Exemple: Name = Foot Description = test => I can change only the description and save it. The name already exists on DB but this is the name who have not changed.

if we change this name by an other exists on DB, ther is an error, because we change the CURRENT name by an other who already exists on db