Changed the structure of our controllers and changed the routes
All "critical" controllers are now on the App/Http/controllers/Admin folder and all "non-critical" controllers are on the normalcontrollers folder. 

This way, it's a better structure and later we can easly add a middleware who will check if the user can or cannot go to the url he asked.