Added functionnality for display sport with court linked or not
On the Edit mehod, we get an array with sport who have one or more court linked and an array who have no court linked.

This way, we can display on the edit view, a dropdown list with two groups. One with the sport who have court linked an another with sport who don't have any court linked.