Changed logic behind schedule
Matches are displayed if they don't have a score set. Before it was based of the current time and match time.
When the page refreshes, all the matches on the page are removed and the max amout the page can have are displayed. (its posible that no matches need updating. the return value will be the exact same matches)